Add files via upload
Web Scraping is the process of using bots to extract data and content from a website.

Library used for scraping data is *BeautifulSoup* :   Beautiful Soup is a Python package that parses HTML and XML documents, including those with malformed markup. It creates a parse tree for documents that can be used to extract data from HTML, which is useful for web scraping. Beautiful Soup also provides simple methods for navigating, searching, and modifying a parse tree in HTML, XML files.